# gateway/api/generated/openapi.json
{"openapi": "3.1.0", "info": {"title": "Dressr", "version": "0.1.0"}, "paths": {"/webhook/transactions": {"post": {"summary": "Handle Transaction Webhook Sync", "description": "Handle a webhook sync.", "operationId": "handle_transaction_webhook_sync_webhook_transactions_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionSyncUpdate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users": {"get": {"tags": ["users"], "summary": "Get Users", "description": "Retrieve users.", "operationId": "get_users_users_get", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/GetUsersQuery"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/User"}, "type": "array", "title": "Response Get Users Users Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/me": {"get": {"tags": ["users"], "summary": "Get Current User", "operationId": "get_current_user_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/transactions": {"get": {"tags": ["transactions"], "summary": "Get Transactions", "description": "Retrieve items.", "operationId": "get_transactions_transactions_get", "parameters": [{"name": "offset", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Offset"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 50, "title": "Limit"}}, {"name": "user_id", "in": "query", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Transaction"}, "title": "Response Get Transactions Transactions Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["transactions"], "summary": "Create Transaction", "description": "Create an transaction.", "operationId": "create_transaction_transactions_post", "security": [{"OAuth2PasswordBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateTransactionRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Transaction"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"tags": ["transactions"], "summary": "Update Transaction", "description": "Create an transaction.", "operationId": "update_transaction_transactions_patch", "security": [{"OAuth2PasswordBearer": []}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateTransactionRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Transaction"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/transactions/refresh": {"post": {"tags": ["transactions"], "summary": "Refresh Transactions", "description": "Refresh transactions.", "operationId": "refresh_transactions_transactions_refresh_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/login/token": {"post": {"summary": "Handle User Login", "description": "OAuth2 compatible token login, get an access token for future requests", "operationId": "handle_user_login_login_token_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserLoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessfulLoginResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/register": {"post": {"summary": "Handle Register User", "description": "Create new user.", "operationId": "handle_register_user_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateUserRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SuccessfulLoginResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CreateTransactionRequest": {"properties": {"currency_code": {"type": "string", "title": "Currency Code", "description": "The code of the currency type", "default": "USD"}, "raw_label": {"type": "string", "title": "Raw Label", "description": "The original description from Plaid"}}, "type": "object", "required": ["raw_label"], "title": "CreateTransactionRequest"}, "CreateUserRequest": {"properties": {}, "type": "object", "title": "CreateUserRequest"}, "GetUsersQuery": {"properties": {"offset": {"type": "integer", "title": "Offset", "description": "How many offset items from the beginning to skip", "default": 0}, "limit": {"type": "integer", "title": "Limit", "description": "The number of items to return with the response", "default": 50}, "user_id": {"type": "string", "title": "User ID", "description": "Filter by this user ID"}}, "type": "object", "required": ["user_id"], "title": "GetUsersQuery", "description": "Query for fetching users."}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "SuccessfulLoginResponse": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}, "user": {"allOf": [{"$ref": "#/components/schemas/User"}], "description": "The user who has just logged in"}}, "type": "object", "required": ["access_token", "token_type", "user"], "title": "SuccessfulLoginResponse"}, "Transaction": {"properties": {"id": {"type": "string", "title": "Id", "description": "The ID of the entity"}, "user_id": {"type": "string", "title": "User Id", "description": "The user id who owns this entity"}, "plaid_transaction_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Plaid Transaction Id", "description": "The ID of the transaction in Plaid"}, "raw_label": {"type": "string", "title": "Raw Label", "description": "The original description from Plaid"}, "label": {"type": "string", "title": "Label", "description": "The label of the transaction"}, "amount": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Amount", "description": "The amount of the transaction (in USD)"}, "currency_code": {"type": "string", "title": "Currency Code", "description": "The code of the currency type"}, "source_account": {"type": "string", "title": "Source Account", "description": "The source account for this transaction"}, "destination_account": {"type": "string", "title": "Destination Account", "description": "The destination account for this transaction"}, "tags": {"items": {"type": "string"}, "type": "array", "title": "Tags", "description": "The tags applied to the transaction"}, "reviewed": {"type": "boolean", "title": "Reviewed", "description": "If the transaction has been reviewed"}}, "type": "object", "required": ["id", "user_id", "plaid_transaction_id", "raw_label", "label", "amount", "currency_code", "source_account", "destination_account", "tags", "reviewed"], "title": "Transaction"}, "TransactionSyncUpdate": {"properties": {"webhook_type": {"type": "string", "title": "Webhook Type", "description": "The type of webhook"}, "webhook_code": {"type": "string", "title": "Webhook Code", "description": "The code of the webhook"}, "item_id": {"type": "string", "title": "Item Id", "description": "The item that has been updated"}, "initial_update_complete": {"type": "boolean", "title": "Initial Update Complete", "description": "If the initial update has finished"}, "historical_update_complete": {"type": "boolean", "title": "Historical Update Complete", "description": "If the historical update has completed"}, "environment": {"type": "string", "title": "Environment", "description": "The environment of the sync"}}, "type": "object", "required": ["webhook_type", "webhook_code", "item_id", "initial_update_complete", "historical_update_complete", "environment"], "title": "TransactionSyncUpdate"}, "UpdateTransactionRequest": {"properties": {"id": {"type": "string", "title": "Id", "description": "The ID of the entity to update"}}, "type": "object", "required": ["id"], "title": "UpdateTransactionRequest"}, "User": {"properties": {"id": {"type": "string", "title": "Id", "description": "The ID of the entity"}, "user_id": {"type": "string", "title": "User Id", "description": "The user id who owns this entity"}, "email": {"type": "string", "title": "Email", "description": "The email of the user"}, "transaction_refresh_interval_minutes": {"type": "integer", "title": "Transaction Refresh Interval Minutes", "description": "The refresh rate of transactions"}, "account_refresh_interval_minutes": {"type": "integer", "title": "Account Refresh Interval Minutes", "description": "The refresh rate of accounts"}}, "type": "object", "required": ["id", "user_id", "email", "transaction_refresh_interval_minutes", "account_refresh_interval_minutes"], "title": "User", "description": "A user in the system."}, "UserLoginRequest": {"properties": {"user_email": {"type": "string", "title": "User Email", "description": "The email for login"}, "password": {"type": "string", "title": "Password", "description": "The password for the user"}}, "type": "object", "required": ["user_email", "password"], "title": "UserLoginRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/login/token"}}}}}}